# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AlgoVisuFSSwagger",
    "version": "v1"
  },
  "paths": {
    "/BudgetCalculator/NaivBudget": {
      "get": {
        "tags": [
          "BudgetCalculator"
        ],
        "operationId": "GetNaivBudget",
        "parameters": [
          {
            "name": "RemainingCapital",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NaivBudgetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NaivBudgetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NaivBudgetDto"
                }
              }
            }
          }
        }
      }
    },
    "/BudgetCalculator/CompleteBudget": {
      "get": {
        "tags": [
          "BudgetCalculator"
        ],
        "operationId": "GetCompleteBudget",
        "parameters": [
          {
            "name": "RemainingCapital",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "MinimumPossibleSpending",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteBudgetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteBudgetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteBudgetDto"
                }
              }
            }
          }
        }
      }
    },
    "/CsvBuilder": {
      "get": {
        "tags": [
          "CsvBuilder"
        ],
        "operationId": "GetBudgetCsv",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/DFS/convert": {
      "post": {
        "tags": [
          "DFS"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MazeModelDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MazeModelDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MazeModelDto"
                }
              }
            }
          }
        }
      }
    },
    "/DFS/solve": {
      "post": {
        "tags": [
          "DFS"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MazeInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MazeInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MazeInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MazeModelDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MazeModelDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MazeModelDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CellGetDto": {
        "type": "object",
        "properties": {
          "posX": {
            "type": "integer",
            "format": "int32"
          },
          "posY": {
            "type": "integer",
            "format": "int32"
          },
          "weight": {
            "type": "number",
            "format": "double"
          },
          "state": {
            "type": "integer",
            "format": "int32"
          },
          "isGoal": {
            "type": "boolean"
          },
          "isStart": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "CompleteBudgetDto": {
        "required": [
          "dailyBudget",
          "dailyBudgetAfterSpendingDailyBudget",
          "dailyBudgetAfterSpendingThousand",
          "dailyBudgetTomorrow",
          "possibleMonthlySavings",
          "remainingDays"
        ],
        "type": "object",
        "properties": {
          "dailyBudget": {
            "type": "number",
            "format": "double"
          },
          "dailyBudgetAfterSpendingDailyBudget": {
            "type": "number",
            "format": "double"
          },
          "dailyBudgetAfterSpendingThousand": {
            "type": "number",
            "format": "double"
          },
          "dailyBudgetTomorrow": {
            "type": "number",
            "format": "double"
          },
          "possibleMonthlySavings": {
            "type": "number",
            "format": "double"
          },
          "remainingDays": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "MazeInputDto": {
        "required": [
          "mazeModel",
          "starCell"
        ],
        "type": "object",
        "properties": {
          "mazeModel": {
            "$ref": "#/components/schemas/MazeModelDto"
          },
          "starCell": {
            "$ref": "#/components/schemas/CellGetDto"
          }
        },
        "additionalProperties": false
      },
      "MazeModelDto": {
        "type": "object",
        "properties": {
          "solvable": {
            "type": "boolean"
          },
          "maze": {
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/CellGetDto"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "NaivBudgetDto": {
        "required": [
          "dailySpending"
        ],
        "type": "object",
        "properties": {
          "dailySpending": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      }
    }
  }
}